backend/intelligence/accuracy.py
import sqlite3
import pandas as pd
import yfinance as yf
import datetime
import logging

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

DB_PATH = "backend/data/gitta.db"

def calculate_accuracy():
    """
    Checks past predictions (Morning Reports) against actual closing prices.
    Updates the 'accuracy' field in the database.
    """
    conn = sqlite3.connect(DB_PATH)
    cursor = conn.cursor()

    # 1. Get past morning reports
    cursor.execute("SELECT date, content FROM daily_reports WHERE type='morning_report'")
    reports = cursor.fetchall()

    total_predictions = 0
    correct_predictions = 0

    for date, content in reports:
        # Parse content (assuming JSON or structured text - simplified here)
        # In a real scenario, we'd parse the JSON content to get symbols
        # For now, let's assume we can extract symbols from a 'symbols' table or similar
        # This is a placeholder logic for the 'Accuracy Engine'
        pass
    
    logger.info("Accuracy calculation complete (Placeholder).")
    conn.close()

if __name__ == "__main__":
    calculate_accuracy()
