backend/requirements.txt
flask
requests
pandas
numpy

beautifulsoup4
praw
arxiv
python-dotenv
redis
celery
prometheus-client
gunicorn
smartapi-python

google-api-python-client
google-auth-httplib2
google-auth-oauthlib
APScheduler==3.10.4
Flask-Caching
Flask-Limiter
Flask-Cors
scikit-learn
prometheus-client
python-dotenv
google-generativeai
openai
groq
gunicorn
